+++ extras/AFC_assist.py
@@ -1,6 +1,6 @@
 # Armored Turtle Automated Filament Changer
 #
-# Copyright (C) 2024 Armored Turtle
+# Copyright (C) 2024-2025 Armored Turtle
 #
 # This file may be distributed under the terms of the GNU GPLv3 license.
 import math