+++ extras/AFC_lane.py
@@ -1,6 +1,6 @@
 # Armored Turtle Automated Filament Changer
 #
-# Copyright (C) 2024 Armored Turtle
+# Copyright (C) 2024-2025 Armored Turtle
 #
 # This file may be distributed under the terms of the GNU GPLv3 license.
 import math
@@ -25,6 +25,7 @@ class AssistActive(Enum):
     YES = 1
     NO = 2
     DYNAMIC = 3
+
 class SpeedMode(Enum):
     LONG = 1
     SHORT = 2