+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2025-05-26]
+### Added
+
+- Added `show_macros` configuration option to the `AFC.cfg` file. This option allows a limited set of internal macros 
+to be displayed in mainsail/fluidd. 
+
 ## [2025-05-25]
 ### Fixed
 - Exclude object bug where klipper would error out with max extrude error after excluding an object and